all buttons except recipe buttons working

+++ app/assets/javascripts/cupboards.js.coffee
@@ -26,24 +26,31 @@ $ ->
 			url: "/cupboards"
 			data: data
 	$(document).on "click", "#delete_from_cupboard_button", ->
+		console.log "HEREREREERE"
 		selected = []
 		$("input[name='cupboard_item']:checked").each ->
 			selected.push ($(this).val())
 		data = items_to_delete: selected
-		$.ajax 
-			type: "DELETE"
-			url: "cupboards/#{selected.first}"
-			data: data
+		if selected ==[]
+			return
+		else
+			$.ajax 
+				type: "DELETE"
+				url: "cupboards/#{selected.first}"
+				data: data
 
 	$(document).on "click", "#delete_from_cupboard_and_add_to_grocery_list_button", ->
 		selected = []
-		console.log "clicked"
 		$("input[name='cupboard_item']:checked").each ->
 			selected.push ($(this).val())
-		data = items_to_add: selected
-		$.ajax
-			type: "POST"
-			url: "grocery_lists/"
+		if selected == []
+			return 
+		else 
+			data = items_to_add: selected
+			$.ajax
+				type: "POST"
+				url: "grocery_lists/"
+				data: data
